+++ hikaricp/tests/conftest.py
@@ -24,8 +24,26 @@ def check(instance):
 
 
 @pytest.fixture()
-def mock_metrics():
-    fixture_file = os.path.join(os.path.dirname(__file__), "fixtures", "metrics.txt")
+def mock_micrometer_metrics():
+    fixture_file = os.path.join(os.path.dirname(__file__), "fixtures", "metrics-micrometer.txt")
+
+    with open(fixture_file, "r") as f:
+        content = f.read()
+
+    with mock.patch(
+        "requests.get",
+        return_value=mock.MagicMock(
+            status_code=200,
+            iter_lines=lambda **kwargs: content.split("\n"),
+            headers={"Content-Type": "text/plain"},
+        ),
+    ):
+        yield
+
+
+@pytest.fixture()
+def mock_prometheus_metrics():
+    fixture_file = os.path.join(os.path.dirname(__file__), "fixtures", "metrics-prometheus.txt")
 
     with open(fixture_file, "r") as f:
         content = f.read()