+++ speedtest/datadog_checks/speedtest/__init__.py
@@ -1,4 +1,4 @@
 from .__about__ import __version__
-from .check import MEGABYTE_TO_MEBIBYTE, SpeedtestCheck
+from .check import BYTE_TO_MEBIBYTE, SpeedtestCheck
 
-__all__ = ['__version__', 'MEGABYTE_TO_MEBIBYTE', 'SpeedtestCheck']
+__all__ = ['__version__', 'BYTE_TO_MEBIBYTE', 'SpeedtestCheck']