+++ hbase_master/CHANGELOG.md
@@ -1,16 +1,22 @@
 # CHANGELOG - HBase master
 
-## 1.0.0 / 2017-05-29
+## 1.1.1 / 2024-02-28
 
-***Added***:
+***Fixed***:
 
-* Adds HBase Master Integration. See [#59][1]
+* Fixes `metrics.yaml` to ensure compatibility with Agent 6/7 by default instead of Agent 5.
 
 ## 1.1.0 / 2020-08-21
 
 ***Fixed***:
 
 * Update configuration to use `metrics.yaml` for JMX. See [#697][2]
+
+## 1.0.0 / 2017-05-29
+
+***Added***:
+
+* Adds HBase Master Integration. See [#59][1]
 
 <!---  --->
 [1]: https://github.com/DataDog/integrations-core/pull/59